refactor: 각 네비게이션의 이벤트를 Header로 위임하고 Header에 싱글톤패턴을 적용하여 하나의 인스턴스만 생성되도록 하여 이벤트가 1번만 등록되도록 처리
