+++ src/components/navigation/UnauthenticatedNavigation.ts
@@ -1,14 +1,4 @@
-import { Router, Routes } from "../../utils/router";
-
 export class UnauthentificatedNavigation {
-  private container: HTMLElement;
-
-  constructor(container: HTMLElement) {
-    this.container = container;
-
-    this.attachEventListeners();
-  }
-
   render() {
     return `
     <nav class="bg-white shadow-md p-2 sticky top-14">
@@ -18,22 +8,5 @@ export class UnauthentificatedNavigation {
         </ul>
       </nav>
       `;
-  }
-
-  attachEventListeners() {
-    this.container.addEventListener("click", (event) => {
-      if (event.target instanceof HTMLAnchorElement) {
-        event.preventDefault();
-
-        const href = extractPath(event.target.href);
-
-        Router.push(href as Routes);
-      }
-
-      function extractPath(url: string) {
-        const index = url.indexOf("/");
-        return index !== -1 ? url.substring(index) : "";
-      }
-    });
   }
 }
